+++ refer-check-plugin/src/main/java/com/ss/android/ugc/bytex/refercheck/log/PinpointProblemAnalyzer.java
@@ -7,20 +7,25 @@ import com.ss.android.ugc.bytex.refercheck.InaccessibleNode;
 import org.dom4j.Document;
 import org.dom4j.Element;
 import org.dom4j.io.SAXReader;
+import org.gradle.api.Project;
 import org.gradle.api.artifacts.Configuration;
+import org.gradle.api.file.FileCollection;
+import org.gradle.api.internal.artifacts.configurations.DefaultConfiguration;
 
 import java.io.File;
+import java.lang.reflect.Field;
+import java.util.Set;
 
 /**
  * Created by tanlehua on 2019/4/15.
  */
-class PinpointProblemAnalyzer {
+public class PinpointProblemAnalyzer {
     private Graph graph;
     private GradleCacheResolver cacheResolver;
     private DependencyGraph dependencyGraph;
     private SAXReader reader = new SAXReader();
 
-    PinpointProblemAnalyzer(Graph graph, GradleCacheResolver cacheResolver, Configuration configuration) {
+    private PinpointProblemAnalyzer(Graph graph, GradleCacheResolver cacheResolver, Configuration configuration) {
         this.graph = graph;
         this.cacheResolver = cacheResolver;
         this.dependencyGraph = new DependencyGraph(configuration);
@@ -85,5 +90,44 @@ class PinpointProblemAnalyzer {
         boolean invalid() {
             return groupId == null || artifactId == null || version == null;
         }
+    }
+
+    public static PinpointProblemAnalyzer getPinpointProblemAnalyzer(Project project, String variantName, Graph graph) {
+        Configuration configuration = null;
+        GradleCacheResolver cacheResolver = null;
+        try {
+            configuration = project.getConfigurations().getByName(String.format("%sRuntimeClasspath", variantName));
+            Set<File> files = null;
+            try {
+                files = configuration.getFiles();
+            } catch (Exception e) {
+                try {
+                    Class<? extends Configuration> configurationClass = DefaultConfiguration.class;
+                    Field intrinsicFiles = configurationClass.getDeclaredField("intrinsicFiles");
+                    if (intrinsicFiles != null) {
+                        intrinsicFiles.setAccessible(true);
+                        FileCollection intrinsicFilesObj = (FileCollection) intrinsicFiles.get(configuration);
+                        Class<?> configurationFileCollectionClass = Class.forName(intrinsicFilesObj.getClass().getName());
+                        Field lenient = configurationFileCollectionClass.getDeclaredField("lenient");
+                        lenient.setAccessible(true);
+                        Object lenientValue = lenient.get(intrinsicFilesObj);
+                        lenient.set(intrinsicFilesObj, true);
+                        files = configuration.getFiles();
+                        lenient.set(intrinsicFilesObj, lenientValue);
+                        lenient.setAccessible(false);
+                    }
+                } catch (Exception e1) {
+                    e1.printStackTrace();
+                }
+            }
+            cacheResolver = new GradleCacheResolver();
+            for (File file : files) {
+                cacheResolver.accept(file);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+        return new PinpointProblemAnalyzer(graph, cacheResolver, configuration);
     }
 }