+++ refer-check-plugin/src/main/java/com/ss/android/ugc/bytex/refercheck/log/TipsProvider.java
@@ -1,0 +1,5 @@
+package com.ss.android.ugc.bytex.refercheck.log;
+
+public interface TipsProvider {
+    String provideFilePathInfo(String fileName);
+}